Get ubuntu codename from node

+++ recipes/setup.rb
@@ -6,7 +6,7 @@ chef_gem "aws-sdk" do
 end
 require "aws-sdk"
 
-node.default["chef_postgres"]["release_apt_codename"] = "xenial"
+node.default["chef_postgres"]["release_apt_codename"] = node["lsb"]["codename"] # || "xenial"
 node.default["chef_postgres"]["version"] = "9.6"
 node.default["chef_postgres"]["workload"] = "oltp"
 # node.default['chef_postgres']['pg_config']['data_drive'] = "/mnt/data"
@@ -81,6 +81,19 @@ template "postgresql.conf" do
                                 pg_node: node["chef_postgres"]["pg_config"]["pg_node"] } } })
 end
 
+# Build this on the master so the standbys have the right settings
+template "recovery_conf.source" do
+  group "postgres"
+  mode "0640"
+  owner "postgres"
+  path "/etc/postgresql/#{version}/main/recovery_conf.source"
+  source "recovery_conf.erb"
+  variables({ config: { username: repl_user,
+                        password: repl_pass,
+                        hostname: node["ec2"]["local_hostname"] } })
+  only_if { node["chef_postgres"]["pg_config"]["pg_node"] == "master" }                                
+end
+
 bash "move_data_directory" do
   action :run
   code <<-EOF_MDD
@@ -116,6 +129,7 @@ bash "create_admin_user" do
   user "postgres"
   code "echo \"CREATE USER #{admin_user} WITH PASSWORD '#{admin_pass}' SUPERUSER CREATEDB CREATEROLE; CREATE DATABASE #{admin_user} OWNER #{admin_user};\" | psql -U postgres -d postgres"
   notifies :run, "ruby_block[log_create_admin]", :before
+  only_if { node["chef_postgres"]["pg_config"]["pg_node"] == "master" } 
 end
 
 # Only run this, if generating the info through the defaults.
@@ -127,6 +141,7 @@ file "record_admin" do
   path "/etc/postgresql/#{version}/main/admin_login"
   action :create
   only_if { admin_is_generated }
+  only_if { node["chef_postgres"]["pg_config"]["pg_node"] == "master" } 
 end
 
 bash "create_repl_user" do
@@ -136,17 +151,7 @@ bash "create_repl_user" do
   EOF_CRU
   action :run
   notifies :run, "ruby_block[log_create_repl]", :before
-end
-
-# Only run this, if generating the info through the defaults.
-file "record_repl" do
-  content "user: #{repl_user} password: #{repl_pass}"
-  group "root"
-  mode "0400"
-  owner "root"
-  path "/etc/postgresql/#{version}/main/repl_login"
-  action :create
-  only_if { repl_is_generated }
+  only_if { node["chef_postgres"]["pg_config"]["pg_node"] == "master" } 
 end
 
 directory "/backups/base_backup" do
@@ -155,16 +160,18 @@ directory "/backups/base_backup" do
   mode "0744"
   recursive true
   notifies :run, "ruby_block[log_backup_directory]", :before
+  only_if { node["chef_postgres"]["pg_config"]["pg_node"] == "master" } 
 end
 
 bash "create_base_backup" do
   code <<-EOF_CBB
   rm -rf /backups/base_backup/*
-  pg_basebackup -d 'host=localhost user=#{repl_user} password=#{repl_pass}' -D /backups/base_backup --xlog-method=stream --write-recovery-conf
+  pg_basebackup -d 'host=localhost user=#{repl_user} password=#{repl_pass}' -D /backups/base_backup --xlog-method=stream 
   tar -czf /backups/base_backup.tgz /backups/base_backup/
   EOF_CBB
   action :run
   notifies :run, "ruby_block[log_create_base_backup]", :before
+  only_if { node["chef_postgres"]["pg_config"]["pg_node"] == "master" } 
 end
 
 ruby_block "s3_upload_backup" do
@@ -173,4 +180,5 @@ ruby_block "s3_upload_backup" do
       { bucket: node["chef_postgres"]["s3"]["bucket"],
         source: "/backups/base_backup.tgz" })
   end
+  only_if { node["chef_postgres"]["pg_config"]["pg_node"] == "master" } 
 end
